:recycle: [#1653] move dupicate code to conftest.py
