+++ performance_test/test_zaken_list.py
@@ -1,42 +1,10 @@
-import time
-
-import jwt
 import pytest
 import requests
 from furl import furl
 
+from .conftest import HEADERS, HEADERS_NON_SUPERUSER, HEADERS_NON_SUPERUSER_MANY_TYPES
+
 BASE_URL = furl("http://localhost:8000/zaken/api/v1/")
-
-
-def generate_token(client_id: str, secret: str) -> str:
-    payload = {
-        "iss": "openzaak",
-        "iat": int(time.time()),
-        "client_id": client_id,
-        "user_id": client_id,
-        "user_representation": client_id,
-    }
-    return jwt.encode(payload, secret, algorithm="HS256")
-
-
-TOKEN_SUPERUSER = generate_token("superuser", "superuser")
-HEADERS = {"Authorization": f"Bearer {TOKEN_SUPERUSER}", "Accept-Crs": "EPSG:4326"}
-
-
-TOKEN_NON_SUPERUSER = generate_token("non_superuser", "non_superuser")
-HEADERS_NON_SUPERUSER = {
-    "Authorization": f"Bearer {TOKEN_NON_SUPERUSER}",
-    "Accept-Crs": "EPSG:4326",
-}
-
-
-TOKEN_NON_SUPERUSER_MANY_TYPES = generate_token(
-    "non_superuser_many_types", "non_superuser_many_types"
-)
-HEADERS_NON_SUPERUSER_MANY_TYPES = {
-    "Authorization": f"Bearer {TOKEN_NON_SUPERUSER_MANY_TYPES}",
-    "Accept-Crs": "EPSG:4326",
-}
 
 
 @pytest.mark.benchmark(max_time=60, min_rounds=5)
