+++ src/openzaak/utils/filterset.py
@@ -16,7 +16,7 @@ class OrderingFilter(_OrderingFilter):
 
         ordering = [self.get_ordering_value(param) for param in value]
         return (
-            qs.model.objects.filter(pk__in=Subquery(qs.values("pk")))
+            qs.model.objects.filter(pk__in=Subquery(qs.only("pk").values("pk")))
             .order_by(*ordering)
             .select_related(*qs.query.select_related)
             .prefetch_related(*qs._prefetch_related_lookups)
